RC3 / MOJMAP, Loom 1.14

+++ src/main/java/fi/dy/masa/tweakeroo/event/WorldLoadListener.java
@@ -3,9 +3,9 @@ package fi.dy.masa.tweakeroo.event;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import javax.annotation.Nullable;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.registry.DynamicRegistryManager;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.core.RegistryAccess;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import fi.dy.masa.malilib.interfaces.IWorldLoadListener;
@@ -26,13 +26,13 @@ import fi.dy.masa.tweakeroo.util.MiscUtils;
 public class WorldLoadListener implements IWorldLoadListener
 {
     @Override
-    public void onWorldLoadImmutable(DynamicRegistryManager.Immutable immutable)
+    public void onWorldLoadImmutable(RegistryAccess.Frozen immutable)
     {
         RenderTweaks.setDynamicRegistryManager(immutable);
     }
 
     @Override
-    public void onWorldLoadPre(@Nullable ClientWorld worldBefore, @Nullable ClientWorld worldAfter, MinecraftClient mc)
+    public void onWorldLoadPre(@Nullable ClientLevel worldBefore, @Nullable ClientLevel worldAfter, Minecraft mc)
     {
         // Always disable the Free Camera mode when leaving the world or switching dimensions
         FeatureToggle.TWEAK_FREE_CAMERA.setBooleanValue(false);
@@ -55,7 +55,7 @@ public class WorldLoadListener implements IWorldLoadListener
     }
 
     @Override
-    public void onWorldLoadPost(@Nullable ClientWorld worldBefore, @Nullable ClientWorld worldAfter, MinecraftClient mc)
+    public void onWorldLoadPost(@Nullable ClientLevel worldBefore, @Nullable ClientLevel worldAfter, Minecraft mc)
     {
         DataManager.getInstance().reset(worldAfter == null);
         EntityDataManager.getInstance().reset(worldAfter == null);
@@ -70,10 +70,10 @@ public class WorldLoadListener implements IWorldLoadListener
 
             // Prevents option value de-sync
             if (FeatureToggle.TWEAK_DARKNESS_VISIBILITY.getBooleanValue() &&
-                mc.options.getDarknessEffectScale().getValue() != Configs.Generic.DARKNESS_SCALE_OVERRIDE_VALUE.getDoubleValue())
+                mc.options.darknessEffectScale().get() != Configs.Generic.DARKNESS_SCALE_OVERRIDE_VALUE.getDoubleValue())
             {
-                Configs.Internal.DARKNESS_SCALE_VALUE_ORIGINAL.setDoubleValue(mc.options.getDarknessEffectScale().getValue());
-                mc.options.getDarknessEffectScale().setValue(Configs.Generic.DARKNESS_SCALE_OVERRIDE_VALUE.getDoubleValue());
+                Configs.Internal.DARKNESS_SCALE_VALUE_ORIGINAL.setDoubleValue(mc.options.darknessEffectScale().get());
+                mc.options.darknessEffectScale().set(Configs.Generic.DARKNESS_SCALE_OVERRIDE_VALUE.getDoubleValue());
             }
         }
 
@@ -111,7 +111,10 @@ public class WorldLoadListener implements IWorldLoadListener
 		JsonObject root = new JsonObject();
 		boolean shouldSave = false;
 
-		FileUtils.delete(file);
+		if (Files.exists(file))
+		{
+			FileUtils.delete(file);
+		}
 
 		if (!CameraPresetManager.getInstance().isEmpty())
 		{
